git-svn-id: https://svn.lcsr.jhu.edu/cisst/branches/2011-12-15-fault@3733 0959eb1a-6956-4f60-947b-6092a4e22c97

+++ cisst/cisstStereoVision/svlCCFileIO.h
@@ -124,7 +124,7 @@ protected:
                 inp.getline(buffer,200,delim);
                 std::istringstream inp2(buffer);
                 while(inp2 >> d) {
-                    data[lineNum][numsRead] = d;
+                    data[lineNum][numsRead] = (float)d;
                     numsRead++;
                 }
             }
